requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.11
sentence-transformers>=3.0.1
PyPDF2>=3.0.1
google-generativeai>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.8.2
requests>=2.32.3
httpx>=0.27.0
